Makes all shapes subclasses to Shape.h
All shapes: Circle.h, Triangle.h, and Square.h, are now subclasses of Shape.h.
This is meant to make it easier for menu selection of which type of shape the user chooses.
Also improves upon contains() function in Circle.h.
All contains() functions still need to be tested properly and added as virtual functions to Shape.h.