Fixes incorrect color error with target
By changing a few numbers in Shape.cpp, this fixes the error that occurred when a user clicked start without selecting a color. By default, the color of the target should be red, but the target was defaulting to green. This simply changes the default constructor to make a default shape red.
Also changes the game's time to 30 seconds instead of 10. 30 seconds should be the normal time. This may be changed back to 10 seconds due to testing purposes.